move progress tracker to models
