+++ src-tauri/src/core/downloads/models.rs
@@ -1,4 +1,6 @@
 use std::collections::HashMap;
+use std::sync::Arc;
+use tokio::sync::Mutex;
 use tokio_util::sync::CancellationToken;
 
 #[derive(Default)]
@@ -28,4 +30,32 @@ pub struct DownloadItem {
 pub struct DownloadEvent {
     pub transferred: u64,
     pub total: u64,
+}
+
+/// Structure to track progress for each file in parallel downloads
+#[derive(Clone)]
+pub struct ProgressTracker {
+    file_progress: Arc<Mutex<HashMap<String, u64>>>,
+    total_size: u64,
+}
+
+impl ProgressTracker {
+    pub fn new(_items: &[DownloadItem], sizes: HashMap<String, u64>) -> Self {
+        let total_size = sizes.values().sum();
+        ProgressTracker {
+            file_progress: Arc::new(Mutex::new(HashMap::new())),
+            total_size,
+        }
+    }
+
+    pub async fn update_progress(&self, file_id: &str, transferred: u64) {
+        let mut progress = self.file_progress.lock().await;
+        progress.insert(file_id.to_string(), transferred);
+    }
+
+    pub async fn get_total_progress(&self) -> (u64, u64) {
+        let progress = self.file_progress.lock().await;
+        let total_transferred: u64 = progress.values().sum();
+        (total_transferred, self.total_size)
+    }
 }
